+++ Lib/stars.hpp
@@ -18,7 +18,7 @@ class Stars : public SpaceObject {
   void Generate(uint32_t seed) {
     srand(seed);
     for (auto& s : stars_) {
-      auto p = Point2d(rand() % width_, rand() % height_);
+      auto p = Point2d((float)(rand() % width_), (float)(rand() % height_));
       s = Star(p, rand() % 4 + 1);
     }
   }
